pyproject.toml
[tool.poetry]
name = "libnss-keycloak"
version = "0.1.0"
description = ""
authors = ["Alex Wright <alexwright1324@gmail.com>"]
# readme = "README.md"
# license = "BSD"
packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
python-keycloak = "^4.2.2"
toml = "^0.10.2"

[tool.poetry.scripts]
libnss-keycloak = "app.service:main"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120